# TweetHere-API-GATEWAY/cmd/docs/swagger.yaml
definitions:
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Tweet Here is a social Media Platform
  title: Go + Gin SocialMedia TweetHereeee
  version: 1.0.0
paths:
...
  /users/addpost:
    post:
      consumes:
      - multipart/form-data
      description: Adds a new tweet with optional images
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: Tweet description
        in: formData
        name: discription
        required: true
        type: string
      - description: Image file(s) to upload
        in: formData
        name: files
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add Tweet
      tags:
      - Tweet
...
  /users/commentpost:
    post:
      consumes:
      - application/json
      description: Adds a comment to a post, optionally as a reply to another comment
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: Post ID
        in: query
        name: id
        required: true
        type: integer
      - description: Comment content
        in: query
        name: comment
        required: true
        type: string
      - description: Parent comment ID (for replying)
        in: query
        name: parentid
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Comment Post
      tags:
      - Tweet
  /users/deletecomment:
    delete:
      consumes:
      - application/json
      description: Deletes a specific comment
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: Comment ID
        in: query
        name: commentid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete Comment
      tags:
      - Tweet
  /users/deletpost:
    delete:
      consumes:
      - application/json
      description: Deletes a user's post
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: Post ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete Post
      tags:
      - Tweet
  /users/editcomment:
    put:
      consumes:
      - application/json
      description: Edits a specific comment
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: Comment ID
        in: query
        name: commentid
        required: true
        type: integer
      - description: Updated comment text
        in: query
        name: comment
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Edit Comment
      tags:
      - Tweet
  /users/editpost:
    patch:
      consumes:
      - application/json
      description: Edits a user's tweet with a new description
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: Tweet ID
        in: query
        name: id
        required: true
        type: integer
      - description: New tweet description
        in: query
        name: discription
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Edit Tweet
      tags:
      - Tweet
...
  /users/getcomments:
    get:
      consumes:
      - application/json
      description: Retrieves all comments for a specific post
      parameters:
      - description: Post ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Comments
      tags:
      - Tweet
  /users/getotherspost:
    get:
      consumes:
      - application/json
      description: Retrieves all posts created by a specific user
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Other User's Posts
      tags:
      - Tweet
  /users/getourpost:
    get:
      consumes:
      - application/json
      description: Retrieves all posts created by the logged-in user
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get User's Posts
      tags:
      - Tweet
...
  /users/likepost:
    post:
      consumes:
      - application/json
      description: Allows a user to like a post
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: Post ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Like Post
      tags:
      - Tweet
...
  /users/savepost:
    post:
      consumes:
      - application/json
      description: Adds a post to bookmarks (saves it)
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: Post ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Save Post
      tags:
      - Tweet
...
  /users/unlikepost:
    post:
      consumes:
      - application/json
      description: Allows a user to unlike a post
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: Post ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Unlike Post
      tags:
      - Tweet
  /users/unsavepost:
    post:
      consumes:
      - application/json
      description: Removes a post from bookmarks (unsaves it)
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: Post ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Unsave Post
      tags:
      - Tweet
...
securityDefinitions:
  bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"